refactor: 불필요한 service layer 삭제
