+++ lib/features/recommended_course/di/recommended_course_providers.dart
@@ -1,7 +1,6 @@
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:http/http.dart' as http;
 import 'package:urban_breeze/core/di/core_providers.dart';
-import 'package:urban_breeze/features/recommended_course/application/services/recommended_course_service.dart';
 import 'package:urban_breeze/features/recommended_course/application/use_cases/get_recommended_course_list_usecase.dart';
 import 'package:urban_breeze/features/recommended_course/data/datasources/recommended_course_remote_datasource.dart';
 import 'package:urban_breeze/features/recommended_course/data/repositories/recommended_course_repository_impl.dart';
@@ -26,21 +25,13 @@ recommendedCourseRepositoryProvider = Provider<RecommendedCourseRepository>((
   return RecommendedCourseRepositoryImpl(remoteDataSource: remoteDataSource);
 });
 
-final Provider<RecommendedCourseService> recommendedCourseServiceProvider =
-    Provider<RecommendedCourseService>((Ref<RecommendedCourseService> ref) {
-      final RecommendedCourseRepository repository = ref.watch(
-        recommendedCourseRepositoryProvider,
-      );
-      return RecommendedCourseService(repository: repository);
-    });
-
 final Provider<GetRecommendedCourseListUseCase>
 getRecommendedCourseListUseCaseProvider =
     Provider<GetRecommendedCourseListUseCase>((
       Ref<GetRecommendedCourseListUseCase> ref,
     ) {
-      final RecommendedCourseService service = ref.watch(
-        recommendedCourseServiceProvider,
+      final RecommendedCourseRepository repository = ref.watch(
+        recommendedCourseRepositoryProvider,
       );
-      return GetRecommendedCourseListUseCase(service: service);
+      return GetRecommendedCourseListUseCase(repository: repository);
     });
